networks with arbitrary structure are now working, temporal connections supported.
genetic mutators are also programmed. mutate is fully functional and
uses a probability matrix to calculate which actions to take (randomly).
sleection algorithm needs a bit of love since it can pick weird stuff
somethimes, but all in all seems to work ok.

crossover has been tested very minimally, but the algorithm is
sound...there just may be bugs in the implementation.

every change to a network is tracked by the networks genome...in other
words, any evolution to the network (new connection, split connection,
remov connection, etc) is scripted into its DNA essentially. this was
the only way i could think of to encode a network. instead of encode the
final state as a gene, encode the steps it took the create the entire
network, step by step (which is a procedural/linear pattern, and easily
genetically modifyable).

this is fucking awesome. once i test crossover and make sure it doesnt
have any horrible issues, im ready to package this up as a library and
use it for my bugs-eating-food test (the usual). we'll see how that
goes. i may want to take a look at the real NNEAT algorithms way of
doing speciation, but for now one population should be fine.

this is awesome.